Fix size_t <==> unsigned int, template std::max function type incompatibility
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@2422 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-qt/src/CSVParser.cpp
@@ -33,7 +33,7 @@ const vector<vector<string> > & CSVParser::parse(const string& fileName, const s
     lines.push_back(line);
   }
   tokens.resize(lines.size());
-  unsigned int maxNumberOfCol = 0;
+  size_t maxNumberOfCol = 0;
   for (int i = 0; i < lines.size(); ++i) {
     tokenize(lines[i], tokens[i], separator, numberOfCol);
     maxNumberOfCol = max(maxNumberOfCol, tokens[i].size());
